requirements.txt
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
certifi==2025.1.31
colorama==0.4.6
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.68.2
#psycopg2-binary>=2.9.3 
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-telegram-bot==21.6
redis==5.2.1
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
